wizz-fare-finder-api/requirements.txt
fastapi==0.115.0
apscheduler
uvicorn[standard]==0.32.0

sqlalchemy==2.0.34
psycopg2-binary==2.9.10

requests==2.32.3
aiohttp==3.10.5
python-dateutil==2.9.0.post0

python-dotenv==1.2.1

